pyproject.toml
[project]
name = "fyp_backend"
version = "v0.1.3"
description = "FYP"
authors = [
    {name = "Sohail"},
]
requires-python = ">=3.10"

dependencies = [
    "fastapi>=0.116",
    "uvicorn[standard]>=0.35",
    "pydantic>=2.11",
    "fastapi-users[sqlalchemy,oauth]>=14.0",
    "python-multipart>=0.0.20",
    "pydantic-settings>=2.10",
    "alembic>=1.15",
    "asyncpg>=0.30",
    "psycopg2-binary>=2.9",
    "celery>=5.3",
    "redis>=6.4",
    "Jinja2>=3.1.6",
    "pillow>=12.0.0",
    "gunicorn>=21.2,<22.0",
    "onnxruntime>=1.23.2",
    "matplotlib>=3.10.7",
    "opencv-python>=4.11.0",
    "numpy>=1.24.3",
    "transformers>=4.55.4",
    "optimum>=2.0.0",
    "optimum[onnxruntime]",
    "huggingface_hub",
    "ultralytics",
    "cloudinary>=1.44.1",
]

[dependency-groups]
prod = [

    "gunicorn>=21.2,<22.0",
]
dev = [
    "pytest>=8.4.1",
    "httpx>=0.28.1",
    "ruff>=0.12.11",
    "mypy>=1.17.1",
    "pre-commit>=3.7.0",
]


[tool.ruff]
line-length = 88
target-version = "py310"
lint.extend-select = ["I"]
fix = true
exclude = [
    ".venv", "my_cahe", "ruff_cache", "*.ipynb"
]

[tool.ruff.lint.isort]
known-first-party = ["app"]
combine-as-imports = true
force-single-line = false

[tool.mypy]
explicit_package_bases = true
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
exclude = [
    "alembic/",
]
